Update the update checker
